Move testers to terrain-planner-benchmarks
