+++ terrain_planner_benchmark/src/test_ompl_dubins_to_circle.cpp
@@ -46,9 +46,9 @@
 #include <visualization_msgs/MarkerArray.h>
 #include <grid_map_ros/GridMapRosConverter.hpp>
 
+#include "terrain_navigation_ros/visualization.h"
 #include "terrain_planner/common.h"
 #include "terrain_planner/terrain_ompl_rrt.h"
-#include "terrain_planner/visualization.h"
 
 double mod2pi(double x) { return x - 2 * M_PI * floor(x * (0.5 / M_PI)); }
